Basic Auth

+++ src/main/java/com/example/oopone/controller/AdminController.java
@@ -14,6 +14,7 @@ import java.math.BigDecimal;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Optional;
 
@@ -42,32 +43,41 @@ public class AdminController {
 
     // Clicking on categories, takes me to all the categories present.
     @GetMapping(value = "/admin/categories")
-    public List<Category> getCat(){
-        return categoryService.getCategories();
+    public List<Category> getCat(@RequestHeader HashMap<String, String> header){
+        String adminPass = "1234";
+        if(header.get("token").equals(adminPass)){return categoryService.getCategories();}
+        else return null;
     }
 
     //Get Category name by Id
     @GetMapping(value = "/admin/category/{id}")
-    public Optional<Category> getCatById(@PathVariable int id){
-        return categoryService.getCategoryById(id);
+    public Optional<Category> getCatById(@RequestHeader HashMap<String, String> header,@PathVariable int id){
+        String adminPass = "1234";
+        if(header.get("token").equals(adminPass)){return  categoryService.getCategoryById(id);}
+        else return null;
     }
 
     //Adding new categories.
     @PostMapping(value = "/admin/categories/add")
-    public Category addCategory(@RequestBody Category category){
-        return categoryService.saveCategory(category);
+    public Category addCategory(@RequestHeader HashMap<String, String> header, @RequestBody Category category){
+        String adminPass = "1234";
+        if(header.get("token").equals(adminPass)){return categoryService.saveCategory(category);}
+        else return null;
     }
 
     //Deleting categories.
     @DeleteMapping("/admin/categories/delete/{id}")
-    public void deleteCategory(@PathVariable int id){
-        categoryService.removeCategory(id);
+    public void deleteCategory(@RequestHeader HashMap<String, String> header, @PathVariable int id){
+        String adminPass = "1234";
+        if(header.get("token").equals(adminPass)){categoryService.removeCategory(id);}
     }
 
     //Updating categories.
     @PutMapping("/admin/categories/update/{id}")
-    public Category updateCategory(@PathVariable int id, @RequestBody Category category){
-        return categoryService.updateCategoryById(id, category);
+    public Category updateCategory(@RequestHeader HashMap<String, String> header, @PathVariable int id, @RequestBody Category category){
+        String adminPass = "1234";
+        if(header.get("token").equals(adminPass)){return categoryService.updateCategoryById(id, category);}
+        else return null;
     }
 
 
@@ -75,44 +85,55 @@ public class AdminController {
 
     //Getting all items.
     @GetMapping("/admin/items")
-    public List<Item> getItems(){
-        return itemService.getItems();
+    public List<Item> getItems(@RequestHeader HashMap<String, String> header){
+        String adminPass = "1234";
+        if(header.get("token").equals(adminPass)){return itemService.getItems();}
+        else return null;
     }
 
     //IN PROCESS.
     @GetMapping("/admin/item/maxoffer")
-    public Item getItemByMaxOffer(){
-        return itemService.getItemWithMaxOffer();
+    public Item getItemByMaxOffer(@RequestHeader HashMap<String, String> header){
+        String adminPass = "1234";
+        if(header.get("token").equals(adminPass)){return itemService.getItemWithMaxOffer();}
+        else return null;
     }
 
 
     @GetMapping("/admin/items/category/{id}")
-    public List<Item> getItemsByCategory(@PathVariable int id){
-        return itemService.getAllItemsByCategoryId(id);
+    public List<Item> getItemsByCategory(@RequestHeader HashMap<String, String> header,@PathVariable int id){
+        String adminPass = "1234";
+        if(header.get("token").equals(adminPass)){return itemService.getAllItemsByCategoryId(id);}
+        else return null;
     }
 
 
     // Adding new items.
     @PostMapping("/admin/items/add/{id}")
-    public Item addItemByCatId(@PathVariable int id , @RequestBody Item item) {
-        return itemService.saveItem(id,item);
+    public Item addItemByCatId(@RequestHeader HashMap<String, String> header,@PathVariable int id , @RequestBody Item item) {
+        String adminPass = "1234";
+        if(header.get("token").equals(adminPass)){return itemService.saveItem(id,item);}
+        else return null;
     }
 
     //Deleting items.
     @DeleteMapping("/admin/items/delete/{id}")
-    public void deleteItem(@PathVariable int id){
-        itemService.removeItemById(id);
+    public void deleteItem(@RequestHeader HashMap<String, String> header,@PathVariable int id){
+        String adminPass = "1234";
+        if(header.get("token").equals(adminPass)){itemService.removeItemById(id);}
     }
 
     //Updating Items.
     @PutMapping("/admin/items/update/{id}")
-    public Item updateItem(@PathVariable int id, @RequestBody Item item){
-        return itemService.updateItemById(id,item);
+    public Item updateItem(@RequestHeader HashMap<String, String> header,@PathVariable int id, @RequestBody Item item){
+        String adminPass = "1234";
+        if(header.get("token").equals(adminPass)){return itemService.updateItemById(id,item);}
+        else return null;
     }
 
     @DeleteMapping("/admin/user/delete/{id}")
-    public void deleteUser(@PathVariable int id){
-        userService.removeUser(id);
+    public void deleteUser(@RequestHeader HashMap<String, String> header,@PathVariable int id){
+        String adminPass = "1234";
+        if(header.get("token").equals(adminPass)){userService.removeUser(id);}
     }
-
 }
